Separate S3 and File IMetsStorage implementations
