+++ src/DigitalPreservation/Storage.Repository.Common/Mets/MetsManager.cs
@@ -1,5 +1,4 @@
-﻿using Amazon.Runtime.Internal.Util;
-using DigitalPreservation.Common.Model;
+﻿using DigitalPreservation.Common.Model;
 using DigitalPreservation.Common.Model.Mets;
 using DigitalPreservation.Common.Model.Results;
 using DigitalPreservation.Common.Model.Transit;
