+++ src/DigitalPreservation/XmlGen.Tests/MetsManagerTests.cs
@@ -8,6 +8,7 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using Moq;
 using Storage.Repository.Common.Mets;
+using Storage.Repository.Common.Mets.StorageImpl;
 
 namespace XmlGen.Tests;
 
@@ -27,7 +28,7 @@ public class MetsManagerTests
         var s3Client = new Mock<IAmazonS3>().Object;
         
         parser = new MetsParser(s3Client, parserLogger);
-        var metsStorage = new MetsStorage(s3Client, parser);
+        var metsStorage = new FileSystemMetsStorage(parser);
         metsManager = new MetsManager(parser, metsStorage);
     }
