+++ src/DigitalPreservation/XmlGen.Tests/MetsManagerWithPremis.cs
@@ -6,6 +6,7 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using Moq;
 using Storage.Repository.Common.Mets;
+using Storage.Repository.Common.Mets.StorageImpl;
 
 namespace XmlGen.Tests;
 
@@ -24,7 +25,7 @@ public class MetsManagerWithPremis
         var parserLogger = factory!.CreateLogger<MetsParser>();
         var s3Client = new Mock<IAmazonS3>().Object;
         parser = new MetsParser(s3Client, parserLogger);
-        var metsStorage = new MetsStorage(s3Client, parser);
+        var metsStorage = new FileSystemMetsStorage(parser);
         metsManager = new MetsManager(parser, metsStorage);
     }
